Added QwtScaleDraw.setFixedSize/fixedSize methods to set the new fixed size option
When drawing text labels, if fixed size option is enabled, the axes
width and height won't change when axis range is changing. On the
contrary, if this option is disabled (default behavior), axes are
drawn in order to optimize layout space and depends on text label
individual sizes.

This option is not implemented in Qwt C++ library: this may be used
either as an optimization (updating plot layout is faster when this
option is enabled) or as an appearance preference (with Qwt default
behavior, the size of axes may change when zooming and/or panning
plot canvas which in some cases may not be desired).